test: split PDF visual source crop boundary coverage
